reorganized integration pipeline
