+++ integration-tests/k8s/pipeline.param
@@ -6,9 +6,9 @@
 # Template File: templates/pipeline.json
 #=========================================================
 NAME="integration"
-PIPELINE_PURPOSE="test"
+PIPELINE_PURPOSE="base"
 GIT_REPO_URL="https://github.com/bcgov/entity.git"
 GIT_REF="master"
 SOURCE_CONTEXT_DIR="integration-tests"
-JENKINS_FILE="Jenkinsfiles/integration-test-pipeline.groovy"
+JENKINS_FILE="Jenkinsfiles/integration-pipeline.groovy"
 WEBHOOK="unknown"
